requirements.txt
black==25.1.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
kiwisolver==1.4.8
matplotlib==3.10.3
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.15.3
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
